user registrations

+++ base/views.py
@@ -5,11 +5,20 @@ from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib import messages
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
+from django.contrib.auth.forms import UserCreationForm
 
 def loginPage(request):
+    
+    page = 'login'
+    
+    if request.user.is_authenticated:
+        return redirect('home')
+    
     if request.method == 'POST':
-        username = request.POST.get('username')
-        password = request.POST.get('password')
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password').lower()
         
         try:
             user = User.objects.get(username=username)
@@ -24,12 +33,30 @@ def loginPage(request):
         else:
             messages.error(request, 'Invalid username or password')
             
-    context = {}
+    context = {'page' : page}
     return render(request, 'base/login_register.html', context)
 
 def logoutUser(request):
     logout(request)
     return redirect('home')
+
+def registerUser(request):
+    page = 'register'
+    form = UserCreationForm()
+    
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False) # Create the user instance but don't save it to the database yet
+            user.username = user.username.lower()  # Modify the user instance
+            user.save() # Save the user instance to the database
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'An error occured during registration!')
+    context = {'page' : page,
+               'form' : form}
+    return render(request, 'base/login_register.html', context)
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -56,6 +83,7 @@ def room(request,pk):
     }
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='/login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -66,9 +94,13 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='/login')
 def updateRoom(request,pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+    
+    if request.user != room.host:
+        return HttpResponse('<h1>You are not authorized to edit this room</h1>')
     
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -79,8 +111,13 @@ def updateRoom(request,pk):
     context = {'form':form}
     return render(request, 'base/room_form.html',context)
 
+@login_required(login_url='/login')
 def deleteRoom(request,pk):
     room = Room.objects.get(id=pk)
+    
+    if request.user != room.host:
+        return HttpResponse('<h1>You are not authorized to edit this room</h1>')
+    
     if request.method == "POST":
         room.delete()
         return redirect('home')
